히스토리/requirements.txt
selenium==4.15.2
Pillow==10.1.0

